Extract JFR symbol lookup (#1378)
