+++ src/index.h
@@ -18,28 +18,43 @@ class Index {
     std::unordered_map<std::string, size_t> _idx_map;
   
   public:
-    Index() = default;
+    Index() {
+        // Index 0 should contain the empty string
+        indexOf("");
+    }
 
     Index(const Index&) = delete;
     Index(Index&&) = delete;
     Index& operator=(const Index&) = delete;
     Index& operator=(Index&&) = delete;
 
+    size_t indexOf(const char* value) {
+        return indexOf(std::string(value));
+    }
+
+    size_t indexOf(const char* value, size_t len) {
+        return indexOf(std::string(value, len));
+    }
+
     size_t indexOf(const std::string& value) {
         return _idx_map.insert({value, _idx_map.size()}).first->second;
+    }
+
+    size_t indexOf(std::string&& value) {
+        return _idx_map.insert({std::move(value), _idx_map.size()}).first->second;
     }
 
     size_t size() const {
         return _idx_map.size();
     }
 
-    void forEachOrdered(const std::function<void(const std::string&)>& consumer) const {
+    void forEachOrdered(const std::function<void(size_t idx, const std::string&)>& consumer) const {
         std::vector<const std::string*> arr(_idx_map.size());
         for (const auto& it : _idx_map) {
             arr[it.second] = &it.first;
         }
         for (size_t idx = 0; idx < size(); ++idx) {
-            consumer(*arr[idx]);
+            consumer(idx, *arr[idx]);
         }
     }
 };
